add MVVM test files
