+++ FunctionConfirmTests/MVVMSubmitViewModelTest.swift
@@ -1,0 +1,28 @@
+//
+//  MVVMSubmitViewModelTest.swift
+//  FunctionConfirmTests
+//
+//  Created by Maki, Yusaku | Mackey | ECID on 2019/01/06.
+//  Copyright © 2019年 牧宥作. All rights reserved.
+//
+
+import XCTest
+
+class MVVMSubmitViewModelTest: XCTestCase {
+
+    override func setUp() {
+        super.setUp()
+    }
+
+    override func tearDown() {
+        super.tearDown()
+    }
+
+    func testExample() {
+    }
+
+    func testPerformanceExample() {
+        self.measure {
+        }
+    }
+}
